requirements.txt
openai
langchain~=0.3.27
pgvector
SQLAlchemy>=2.0
psycopg2-binary
arize-phoenix
pandas~=2.3.1
pydantic~=2.11.7
tenacity
PyPDF2~=3.0.1
fastapi~=0.116.1
uvicorn~=0.35.0
requests~=2.32.4